feat: complete CRUD for categories and items endpoints + improve error handling for malformed json request bodies

+++ src/main/java/br/com/luangenro/menu/manager/domain/dto/ItemResponse.java
@@ -1,6 +1,7 @@
 package br.com.luangenro.menu.manager.domain.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import lombok.Builder;
 
 /**
@@ -17,4 +18,4 @@ public record ItemResponse(
     @Schema(description = "Unique ID of the item.", example = "101") Integer id,
     @Schema(description = "Name of the item.", example = "Coca-Cola") String name,
     @Schema(description = "Description of the item.", example = "350ml can") String description,
-    @Schema(description = "Price of the item.", example = "5.50") Double price) {}
+    @Schema(description = "Price of the item.", example = "5.50") BigDecimal price) {}
